Added a basic physical frame allocator.
The frame allocator simply takes the list of useable memory, and allocates frames, one at a time, from the head of the list.
There is no way to free memory using this method.

Later, this shall be used as a "runt mode", as opposed to the "full mode" which shall maintain a list of memory regions that
are free, [start, end) to allocate from in a vector that can grow and shrink as necessary.

The runt mode shall be used to bring up a very basic kernel memory allocator, which can then be used to allocate this vector.
We need a memory manager in order to make the memory allocator work.