Add a kernel heap.
A page allocator has been created; page allocators are directly tied to page tables, and are used to allocate page ranges
for use, and free them again afterwards.

Right now, the page allocator does not deallocate pages, and only runs in runt mode, in the same way as the frame allocator.
Runt mode is a special mode reserved for pre-heap-init operation, in which heap allocations cannot occur.

Furthermore, a 100KiB heap has been set up for the kernel to use, and has been tested.