Added full mode enablement for the Kernel page allocator.
The algorithm itself needs some improvement. It iterates over the kernel page table at call time, building a map of free regions iteratively.

For some reason this takes a long time, and takes a lot of heap - more than it should. This can almost certainly be improved.

Furthermore, an outstanding todo is to make actual use of this memory map when allocating, which is an algorithm yet unwritten.