Added the ability to map MMIO devices without needing to check for existence of RAM in that region. MMIO devices will be allocate from contiguous pieces of physical memory from a defined starting address.
The same function can eventually be used to allocate contiguous regions of usable RAM, for example if needed for DMA. This has not
yet been implemented.

It is probably also worth implementing a check to ensure the contiguous MMIO request does not overlap with, or conflict in any way
with, usable RAM, allocated RAM, or certain types of reserved memory such as the kernel itself. This has not yet been done.