Initialise the BSP APIC.
It is possible, nay likely, that this same code can be re-used for APs, but for now we don't support them, and that's an assumption,
so we check for BSP and force it.

Initialise the APIC in X2 mode. Non-X2 APICs are not supported, though easily could be; neither are APIC-less systems, though
could probably easily be as well.

We should probably re-map and disable the PIC explicitly, though that's not done yet either.