Initialise ACPI tables.
Use the ACPI tables to determine whether or not a legacy PIC is present and needing to be remapped. This is yet to happen.

Also provides an ACPI memory mapper, which maps in memory as the ACPI subsystem needs it. This is currently crude, and does not
yet check whether the requested memory is mapped in.