Initialise the I/O APIC, configure it with legacy IRQs taking account of any present overrides. Enable hardware interrupts.
A few issues in the memory manager (and elsewehere) were found and corrected in the coures of getting the I/O APIC working.
This is the first real use in anger of the memory manager code, so this was to be expected.

Some improvements to be made in the future include moving interrupts into its own folder, that can then be broken down into
Local APIC, I/O APIC, and IDT files, as interrupts is now getting a little unwieldy. In addition, the system should be put
into APIC mode, but this requires AML support, which isn't yet present.