Added a driver model, that scans through the ACPI tables to find relavent driver IDs and attempts to load them, if a driver exists.
Create a stub HPET driver, that currently doesn't do anything other than setting up a dummy device.

In future, we will need to install drivers from other places too - notably, but probably not limited to, PCI(e) and USB. The model will need extending to support this.
Further, we will need support (a) for virtual devices, and (b) for device types. In this case, the HPET is a system level device for the OS only, but others will be
character devices, block devices, or something more complex still, with appropriate methods for reads, writes, ioctls, etc.