Add support for HPET interrupts.
In this commit, we added the abiltiy to enable/disable at will I/O APIC GSIs, and add handler functions to the IDT. Right now, this
is not very generic, and there are lots of improvements that can be made, but it ges us off the ground.

Tested this by having a rolling timer - HPET counter 0 triggers counter 1 triggers counter 2, and so on, til it wraps around again.