List PCI devices on the screen.
Refactor devices to support buses. A bus can be enumerated into devices (for now, only ACPI devices, we're fairly restrictive on that). Add a new driver for PCI, which can scan through the PCI bus.

At present, we don't support PCIe. We also don't save any system resources given in _CRS, which we probably should. Right now, we don't yet create, or enumerate, the PCI bus as a device bus, we simply
scan through it, printing any IDs we find.