Detect and enumerate IDE hard disks.
Right now, only supports disks in compatibility mode (legacy, ISA port addresses), and doesn't support more than 2 IDE ports.
Furthermore, if, on init, a drive encounters an error, we don't bother continuing to enumerate it.

At this stage, a way of sleeping is needed, to allow the commands to be sent with enough of a delay. We don't do that yet.