Added the ability to read files from FAT, from the root directory.
Right now, the code to manage FAT isn't very generic or portable, there are
a lot of potential tidy-ups that can happen over time, and almost certainly
will. Including, but not limited to, caching directories.

Furthermore, fix a bug in which the IDE interface was driven in CHS mode,
not LBA mode.