Create the initial execution stack, before transferring control to a program.
There are multiple limitations with this approach. Firstly, it happens every time we transfer control, which is obviously
wrong. Secondly, it locks the process list, then jumps away never to be seen again, meaning it will never be unlocked.
Lastly, aux vectors should be in an array, rather than hard coded like this.