Load ld.so.
At present we do not use ld.so, nor do we set up relocations, but it is successfully present in memory at this point with no issues.
This entails adding code to be able to load in dynamic shared objects in their entirety, including being able to allocate a block
of memory for the entire ELF file, which may or may not be at some fixed address, or a random dynamic address decided by the allocator.