Implement infrastructure facilitating fork() and process context switch.
The scheduler now registers a (one shot) timer callback with the HPET at time of fork(), facilitating a singular context switch. This is because, for now,
periodic timers are not supported by the HPET.

Updates to syscall entry and interrupt handler entry have been made, to be scheduler aware as well as preserving process state.

A full copy of the parent process is made, then run at the appropriate time.

Problematically, this currently does not clone the address space, merely creating a new stack in the existing address space. This causes stack confusion, meaning
the execution of a LEAVE instruction throws everything off. The result is that we cannot currently leave libc, though scheduling does work.

Todo in the future: clone the entire address space, with the stack either copied to a new set of pages, or marked as COW.