Discover initial device on UHCI ports, using a GET_CONFIGURATION packet.
This initialises a DMA Arena allocator (new), which allows sub-page allocations within a DMA region. In turn, it then initialises
transfer descriptors for SETUP, IN, and OUT, to obtain a DEVICE descriptor.

In future, the DMA Arena allocator can be made to allocate further DMA pages if necessary. In addition, error checking needs to be
properly introduced to USB transfer, to check each of the transfer descriptors' statuses.

PCI interrupt routing would be very helpful here, and the point is quickly being reached wherein the lack thereof is becoming a
limitation.

The next step, however, is to use the so-called "short descriptor" to request a full configuration,  which can consist of multiple
descriptors. This can then be used to fully initialise the port.