Start integration with uACPI.
Rust acpi/aml crates have been stretched to their capabilities; they are not capable of putting the machine into ACPI mode and routing GPEs, there is a known bug in the PCI interrupt
routing table parser, and so forth. Though these problems can be fixed with patches upstream, getting the dependency to a point of maturity is a large undertaking, and uACPI is well
tested, well understood, and battle hardened.

Right now, this means going backwards in order to go forwards. uACPI does not yet run. There are a lot of essential kernel helper functions yet to be written. At this stage, it builds
successfully and prints a banner with version number, and that's about it.