Enumerate ACPI devices through uACPI. This completes the uACPI migration.
This exposes, in its turn, an issue with the current driver setup, in that there is no way at current to control which devices
come up, and in what order. Should a driver, for example, need to make use of the HPET to manage timeouts, and the HPET happens
to be the last device enumerated, that won't (at present), work. There needs to be a way to control driver dependencies.