Introduce a keyboard device. The keyboard device first sets up an interrupt in transfer with the HCI. Add interrupt in handling to UHCI.
Currently, there is absolutely no way for the HCI, upon receiving an interrupt, to signal to the attached device that there is an interrupt.
This will need to be added presently, maybe using callbacks, or alternatively, by using a new driver model method.

Further, the UHCI schedule isn't currently much of a schedule. This needs work. At present, we expect that whatever the latest "thing" is is the
only "thing" in that particular frame link. This is obviously not correct. We need to add the ability to properly join together Isochronous, Interrupt,
and Bulk/Control tranfsers, according to the correct set of rules.

The API is also starting to get slightly messy, as is the "transfer()" method, which needs breaking up into smaller sections. This should be
somewhat straight-forward.

In addition, interrupt transfers need resetting after they fire, to allow the USB device to generate another interrupt. This should be as simple as
resetting status, and inverting toggle.

Lastly, at present, UHCI interrupt handler does not reset the interrupt status, causing interrupts to fire indefinitely.