Decode keypresses, and route appropriately. Deduplicate keypresses when held down simultaneously.
Currently, this cannot handle key rollover, and only outputs whatever the latest key is. In a console setting, this is perfectly acceptable for now. Furthermore,
it decodes neither modifier nor non-ASCII keys.

At present this is simply output to the screen in a somewhat simple fashion. This needs to be routed through to the console device, to be picked up by user software.