Wire through keyboard to console. This requires a significant rewrite of the way read() is implemented, as well as the implementation of ioctl().
Read now passes around Bytes buffers, as opposed to raw pointers, which is much, much better for safety. The way sizes are computed has been reworked, and is now
closer to what the Posix standard expects.

Keyboard input still does not work, for two, related reasons: firstly, if no data is available, Posix expects a tty to return -EAGAIN, rather than 0 bytes. Secondly,
a tty should be in Canonical mode, unless set with O_NONBLOCK - meaning, in practise, that the console should return in units of a line (up to a newline). If no
newline is present, return nothing, even if data is in the buffer.