Return console information to the user process.
At the moment, this doesn't properly block, which causes issues. This assumes, without checking, O_NONBLOCKING. This
causes scanf() to simply return an empty string.

Next step is to properly block usermode processes.

Furthermore, this highlights an issue in which the keyboard slows down with each keypress, and eventually stops
responding. This indicates that the UHCI transactions aren't being properly reset - related to the double interrupt?