Move interrupts to return via the scheduler.
This means that interrupts must never fire outside of a scheduled context. This includes: never in a syscall handler, never in an interrupt handler, and never
in startup code. Or, put another way, the only time an interrupt should happen is either in user code, or in a kthread (once we support kthreads, which we
right now do not).