Refactor to move process into its own module. This allows for clearer ownership, as well as moving all sorts of irrelavent things (open_fd()) out of the scheduler.
This refactoring is as yet incomplete, this is an intermediate milestone. The next step is to start converting everything to getting an Arc to the current process,
and use it directly, rather than the grab-bag of functions in the scheduler right now. As this is refactored, these random assorted functions will be removed.