Make the kernel allocator kernel only.
This necessitated moving a lot of things around, as the kernel allocator was used for all sorts of user allocs. It also breaks a few things, notably getting
buffers from kernel into user spaces, which will be addressed in due course.

Doing this restores the kernel to working, up to the point where it needs to actually activate a user page table.