Refactor VFS to be more production ready, with a more robust VNode -> FileHandle design. This allows for multiple mounts, will later extend to symlinks, and offers better pipe/device semantics.
This was undertaken to build towards a better console setup, in which a graphics buffer is managed by a userspace program, which uses /dev/ptmx to open a master-side
console, creating /dev/ttyX. At present, devices can have only one path attached, which in turn doesn't differentiate its handle; this fixes that.

Currently console still isn't set up for that, but that will come in time.

This has resulted in a large fault, in that read() implicitly happens on the kernel heap. Some debugging output is present in this commit, helping to identify that.
In a future commit, Bytes/BytesMut will be replaced with a page-backed variant, for use within filesystem and block device code.

The new read() for FAT12/16 isn't as peformant, as it no longer does block aggregation. This can easily be added back in, however, and is the subject of a future
commit.